fix small things
